Move extract_content_db to i18n.base.
