+++ kalite/distributed/management/commands/retrievecontentpack.py
@@ -1,7 +1,6 @@
 import os
 import shutil
 import tempfile
-import urllib
 import zipfile
 
 from optparse import make_option
@@ -15,8 +14,8 @@ from fle_utils.general import ensure_dir
 
 from kalite.contentload import settings as content_settings
 from kalite.i18n.base import lcode_to_django_lang, get_po_filepath, get_locale_path, \
-    download_content_pack, update_jsi18n_file, get_srt_path as get_subtitle_path
-from kalite.topic_tools import settings
+    download_content_pack, update_jsi18n_file, get_srt_path as get_subtitle_path, \
+    extract_content_db
 from kalite.updates.management.commands.classes import UpdatesStaticCommand
 
 logging = django_settings.LOG
@@ -145,27 +144,6 @@ def extract_catalog_files(zf, lang):
         logging.debug("writing to %s" % mopath)
         with open(mopath, "wb") as djangomof:
             shutil.copyfileobj(zipmof, djangomof)
-
-
-def extract_content_db(zf, lang, is_template=False):
-    """
-    :param: as_template: Extracts the result to the template destination,
-                         intended for source distribution
-    """
-    if not is_template:
-        content_db_path = settings.CONTENT_DATABASE_PATH.format(
-            channel=settings.CHANNEL,
-            language=lang,
-        )
-    else:
-        content_db_path = settings.CONTENT_DATABASE_TEMPLATE_PATH.format(
-            channel=settings.CHANNEL,
-            language=lang,
-        )
-
-    with open(content_db_path, "wb") as f:
-        dbfobj = zf.open("content.db")
-        shutil.copyfileobj(dbfobj, f)
 
 
 def extract_subtitles(zf, lang):
